network-config-backup/requirements.txt
netmiko>=4.2.0
pyyaml>=6.0
jinja2>=3.1.0
click>=8.1.0
colorama>=0.4.6
prettytable>=3.9.0
apscheduler>=3.10.0
flask>=3.0.0
